[Pipeline] Introduce pipelines to properly rotate channels
we want to limit the number of processor/senders to avoid using
too many resources on host. We previously would have a fixed number,
and assign one of those pipelines to each source. This would create
imbalances.
With this commit, we trully round robin pipelines, so we avoid creating
more bias than there already is (each source being different)