[Docker] Fix issue that would submit log twice on restart
For docker sources.
We store the time of the last log we sent, and start from that date.
Problem is that when we recover tailing, we start from that date
but it's inclusive, so the last log line is returned and sent twice.
A quick workaround is to add 1ns to the last time, so it becomes
exclusive. We can assume that we won't have two log lines with the
same timestamp (nanosecond precision), so we won't skip any log line
with this behavior