[Config] Hostname - don't set hostname to empty string
When the hostname in config is set to `""`, agent6 discards that
value and recomputs it. Log-agent doesn't behave the same way, and
as a result it is confusing.
Let's instead match agent6's behavior.

This is specially important, as we have a script that converts
agent5 config to agent6 config, which defaults the hostname to `""`.