[Processing] Allow overrides of timestamp, severity, tags
Currently we're submitting all logs with INFO severity, current time
as the timestamp, and integration tags.
We'd like to make this more flexible, and let each integration
override it if needed.
To do so, let's make the RFC5424 message we generate a bit more
flexible, without changing the default behavior